Refactor component to remove dropdown and use static link
The component has been simplified by removing the dropdown functionality and replacing it with a static link that displays the company information. The `useState` hook has been removed as the company information is now a constant object. The `Link` component from `next/link` is used to make the company information clickable, redirecting to the home page. Unused imports have been cleaned up accordingly.