Replace hardcoded user validation with API call
The code has been modified to replace the hardcoded user validation with an API call to validate credentials. The `authorize` function now sends a POST request to an API endpoint specified by the `NOTIFY_API_URL` environment variable. The credentials are validated using `loginSchema.parse(credentials)` and then sent in the body of the POST request. If the API response is successful (`response.ok`), the response data is returned. If the API response is not successful, `null` is returned. The previous hardcoded check for a user with username "admin" and password "admin123" has been removed.