Add Login method to UserMastersController, update LoginModel
A new Login method has been added to the UserMastersController class.
This method is an HTTP POST endpoint that takes a LoginModel object as
input and returns a UserMaster object if a user with the specified
username exists in the database.

The LoginModel class has been modified. The properties UserEmail and
Password have been replaced with username and password, respectively.