Refactor ProviderMastersController and update models
Updated ProviderMastersController to include a new namespace import and modified the return types of the `Get` methods to use `ProviderResponseModel`. Implemented `GroupJoin` with `UserTokens` in both `Get` methods to include token information in the response. Added a new property `ProviderName` to `ProviderModel` and introduced a new class `ProviderResponseModel` with relevant properties.